@@ -4,7 +4,6 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"os"
 	"path"
 	"strconv"
 	"strings"
@@ -18,10 +17,11 @@ import (
 	"github.com/Mrs4s/go-cqhttp/global"
 	"github.com/Mrs4s/go-cqhttp/internal/base"
 	"github.com/Mrs4s/go-cqhttp/internal/cache"
+	"github.com/Mrs4s/go-cqhttp/internal/download"
 )
 
 // ToFormattedMessage 将给定[]message.IMessageElement转换为通过coolq.SetMessageFormat所定义的消息上报格式
-func ToFormattedMessage(e []message.IMessageElement, source message.Source) (r interface{}) {
+func ToFormattedMessage(e []message.IMessageElement, source message.Source) (r any) {
 	if base.PostFormat == "string" {
 		r = toStringMessage(e, source)
 	} else if base.PostFormat == "array" {
@@ -61,7 +61,11 @@ func (ev *event) MarshalJSON() ([]byte, error) {
 		fmt.Fprintf(buf, `,"sub_type":"%s"`, ev.SubType)
 	}
 	for k, v := range ev.Others {
-		v, _ := json.Marshal(v)
+		v, err := json.Marshal(v)
+		if err != nil {
+			log.Warnf("marshal message payload error: %v", err)
+			return nil, err
+		}
 		fmt.Fprintf(buf, `,"%s":%s`, k, v)
 	}
 	buf.WriteByte('}')
@@ -132,7 +136,7 @@ func (bot *CQBot) groupMessageEvent(c *client.QQClient, m *message.GroupMessage)
 	bot.dispatch(gm)
 }
 
-func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEvent) {
+func (bot *CQBot) tempMessageEvent(_ *client.QQClient, e *client.TempMessageEvent) {
 	m := e.Message
 	bot.checkMedia(m.Elements, m.Sender.Uin)
 	source := message.Source{
@@ -140,7 +144,10 @@ func (bot *CQBot) tempMessageEvent(c *client.QQClient, e *client.TempMessageEven
 		PrimaryID:  e.Session.Sender,
 	}
 	cqm := toStringMessage(m.Elements, source)
-	bot.tempSessionCache.Store(m.Sender.Uin, e.Session)
+	if base.AllowTempSession {
+		bot.tempSessionCache.Store(m.Sender.Uin, e.Session)
+	}
+
 	id := m.Id
 	// todo(Mrs4s)
 	// if bot.db != nil { // nolint
@@ -468,6 +475,9 @@ func (bot *CQBot) offlineFileEvent(c *client.QQClient, e *client.OfflineFileEven
 }
 
 func (bot *CQBot) joinGroupEvent(c *client.QQClient, group *client.GroupInfo) {
+	if group == nil {
+		return
+	}
 	log.Infof("Bot进入了群 %v.", formatGroupName(group))
 	bot.dispatch(bot.groupIncrease(group.Code, 0, c.Uin))
 }
@@ -481,7 +491,7 @@ func (bot *CQBot) leaveGroupEvent(c *client.QQClient, e *client.GroupLeaveEvent)
 	bot.dispatch(bot.groupDecrease(e.Group.Code, c.Uin, e.Operator))
 }
 
-func (bot *CQBot) memberPermissionChangedEvent(c *client.QQClient, e *client.MemberPermissionChangedEvent) {
+func (bot *CQBot) memberPermissionChangedEvent(_ *client.QQClient, e *client.MemberPermissionChangedEvent) {
 	st := "unset"
 	if e.NewPermission == client.Administrator {
 		st = "set"
@@ -492,7 +502,7 @@ func (bot *CQBot) memberPermissionChangedEvent(c *client.QQClient, e *client.Mem
 	})
 }
 
-func (bot *CQBot) memberCardUpdatedEvent(c *client.QQClient, e *client.MemberCardUpdatedEvent) {
+func (bot *CQBot) memberCardUpdatedEvent(_ *client.QQClient, e *client.MemberCardUpdatedEvent) {
 	log.Infof("群 %v 的 %v 更新了名片 %v -> %v", formatGroupName(e.Group), formatMemberName(e.Member), e.OldCard, e.Member.CardName)
 	bot.dispatchEvent("notice/group_card", global.MSG{
 		"group_id": e.Group.Code,
@@ -516,7 +526,7 @@ func (bot *CQBot) memberLeaveEvent(_ *client.QQClient, e *client.MemberLeaveGrou
 	bot.dispatch(bot.groupDecrease(e.Group.Code, e.Member.Uin, e.Operator))
 }
 
-func (bot *CQBot) friendRequestEvent(c *client.QQClient, e *client.NewFriendRequest) {
+func (bot *CQBot) friendRequestEvent(_ *client.QQClient, e *client.NewFriendRequest) {
 	log.Infof("收到来自 %v(%v) 的好友请求: %v", e.RequesterNick, e.RequesterUin, e.Message)
 	flag := strconv.FormatInt(e.RequestId, 10)
 	bot.friendReqCache.Store(flag, e)
@@ -527,7 +537,7 @@ func (bot *CQBot) friendRequestEvent(c *client.QQClient, e *client.NewFriendRequ
 	})
 }
 
-func (bot *CQBot) friendAddedEvent(c *client.QQClient, e *client.NewFriendEvent) {
+func (bot *CQBot) friendAddedEvent(_ *client.QQClient, e *client.NewFriendEvent) {
 	log.Infof("添加了新好友: %v(%v)", e.Friend.Nickname, e.Friend.Uin)
 	bot.tempSessionCache.Delete(e.Friend.Uin)
 	bot.dispatchEvent("notice/friend_add", global.MSG{
@@ -535,29 +545,31 @@ func (bot *CQBot) friendAddedEvent(c *client.QQClient, e *client.NewFriendEvent)
 	})
 }
 
-func (bot *CQBot) groupInvitedEvent(c *client.QQClient, e *client.GroupInvitedRequest) {
+func (bot *CQBot) groupInvitedEvent(_ *client.QQClient, e *client.GroupInvitedRequest) {
 	log.Infof("收到来自群 %v(%v) 内用户 %v(%v) 的加群邀请.", e.GroupName, e.GroupCode, e.InvitorNick, e.InvitorUin)
 	flag := strconv.FormatInt(e.RequestId, 10)
 	bot.dispatchEvent("request/group/invite", global.MSG{
-		"group_id": e.GroupCode,
-		"user_id":  e.InvitorUin,
-		"comment":  "",
-		"flag":     flag,
+		"group_id":   e.GroupCode,
+		"user_id":    e.InvitorUin,
+		"invitor_id": 0,
+		"comment":    "",
+		"flag":       flag,
 	})
 }
 
-func (bot *CQBot) groupJoinReqEvent(c *client.QQClient, e *client.UserJoinGroupRequest) {
+func (bot *CQBot) groupJoinReqEvent(_ *client.QQClient, e *client.UserJoinGroupRequest) {
 	log.Infof("群 %v(%v) 收到来自用户 %v(%v) 的加群请求.", e.GroupName, e.GroupCode, e.RequesterNick, e.RequesterUin)
 	flag := strconv.FormatInt(e.RequestId, 10)
 	bot.dispatchEvent("request/group/add", global.MSG{
-		"group_id": e.GroupCode,
-		"user_id":  e.RequesterUin,
-		"comment":  e.Message,
-		"flag":     flag,
+		"group_id":   e.GroupCode,
+		"user_id":    e.RequesterUin,
+		"invitor_id": e.ActionUin,
+		"comment":    e.Message,
+		"flag":       flag,
 	})
 }
 
-func (bot *CQBot) otherClientStatusChangedEvent(c *client.QQClient, e *client.OtherClientStatusChangedEvent) {
+func (bot *CQBot) otherClientStatusChangedEvent(_ *client.QQClient, e *client.OtherClientStatusChangedEvent) {
 	if e.Online {
 		log.Infof("Bot 账号在客户端 %v (%v) 登录.", e.Client.DeviceName, e.Client.DeviceKind)
 	} else {
@@ -666,7 +678,8 @@ func (bot *CQBot) checkMedia(e []message.IMessageElement, sourceID int64) {
 			filename := hex.EncodeToString(i.Md5) + ".image"
 			cache.Image.Insert(i.Md5, data)
 			if i.Url != "" && !global.PathExists(path.Join(global.ImagePath, "guild-images", filename)) {
-				if err := global.DownloadFile(i.Url, path.Join(global.ImagePath, "guild-images", filename), -1, nil); err != nil {
+				r := download.Request{URL: i.Url}
+				if err := r.WriteToFile(path.Join(global.ImagePath, "guild-images", filename)); err != nil {
 					log.Warnf("下载频道图片时出现错误: %v", err)
 				}
 			}
@@ -684,12 +697,11 @@ func (bot *CQBot) checkMedia(e []message.IMessageElement, sourceID int64) {
 			i.Name = strings.ReplaceAll(i.Name, "{", "")
 			i.Name = strings.ReplaceAll(i.Name, "}", "")
 			if !global.PathExists(path.Join(global.VoicePath, i.Name)) {
-				b, err := global.GetBytes(i.Url)
+				err := download.Request{URL: i.Url}.WriteToFile(path.Join(global.VoicePath, i.Name))
 				if err != nil {
 					log.Warnf("语音文件 %v 下载失败: %v", i.Name, err)
 					continue
 				}
-				_ = os.WriteFile(path.Join(global.VoicePath, i.Name), b, 0o644)
 			}
 		case *message.ShortVideoElement:
 			data := binary.NewWriterF(func(w *binary.Writer) {
