@@ -14,9 +14,12 @@ import (
 
 	"github.com/Mrs4s/MiraiGo/binary"
 	"github.com/Mrs4s/MiraiGo/client"
+	"github.com/Mrs4s/MiraiGo/wrapper"
 	para "github.com/fumiama/go-hide-param"
 	rotatelogs "github.com/lestrrat-go/file-rotatelogs"
+	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+	"github.com/tidwall/gjson"
 	"golang.org/x/crypto/pbkdf2"
 	"golang.org/x/term"
 
@@ -26,6 +29,7 @@ import (
 	"github.com/Mrs4s/go-cqhttp/global/terminal"
 	"github.com/Mrs4s/go-cqhttp/internal/base"
 	"github.com/Mrs4s/go-cqhttp/internal/cache"
+	"github.com/Mrs4s/go-cqhttp/internal/download"
 	"github.com/Mrs4s/go-cqhttp/internal/selfdiagnosis"
 	"github.com/Mrs4s/go-cqhttp/internal/selfupdate"
 	"github.com/Mrs4s/go-cqhttp/modules/servers"
@@ -41,8 +45,12 @@ var allowStatus = [...]client.UserOnlineStatus{
 	client.StatusGaming, client.StatusVacationing, client.StatusWatchingTV, client.StatusFitness,
 }
 
-// Main 启动主程序
-func Main() {
+// InitBase 解析参数并检测
+//
+//	如果在 windows 下双击打开了程序，程序将在此函数释出脚本后终止；
+//	如果传入 -h 参数，程序将打印帮助后终止；
+//	如果传入 -d 参数，程序将在启动 daemon 后终止。
+func InitBase() {
 	base.Parse()
 	if !base.FastStart && terminal.RunningByDoubleClick() {
 		err := terminal.NoMoreDoubleClick()
@@ -50,18 +58,25 @@ func Main() {
 			log.Errorf("遇到错误: %v", err)
 			time.Sleep(time.Second * 5)
 		}
-		return
+		os.Exit(0)
 	}
 	switch {
 	case base.LittleH:
 		base.Help()
 	case base.LittleD:
 		server.Daemon()
-	case base.LittleWD != "":
-		base.ResetWorkingDir()
+	}
+	if base.LittleWD != "" {
+		err := os.Chdir(base.LittleWD)
+		if err != nil {
+			log.Fatalf("重置工作目录时出现错误: %v", err)
+		}
 	}
 	base.Init()
+}
 
+// PrepareData 准备 log, 缓存, 数据库, 必须在 InitBase 之后执行
+func PrepareData() {
 	rotateOptions := []rotatelogs.Option{
 		rotatelogs.WithRotationTime(time.Hour * 24),
 	}
@@ -91,13 +106,17 @@ func Main() {
 	mkCacheDir(global.VideoPath, "视频")
 	mkCacheDir(global.CachePath, "发送图片")
 	mkCacheDir(path.Join(global.ImagePath, "guild-images"), "频道图片缓存")
+	mkCacheDir(global.VersionsPath, "版本缓存")
 	cache.Init()
 
 	db.Init()
 	if err := db.Open(); err != nil {
 		log.Fatalf("打开数据库失败: %v", err)
 	}
+}
 
+// LoginInteract 登录交互, 可能需要键盘输入, 必须在 InitBase, PrepareData 之后执行
+func LoginInteract() {
 	var byteKey []byte
 	arg := os.Args
 	if len(arg) > 1 {
@@ -131,69 +150,74 @@ func Main() {
 	if base.Debug {
 		log.SetLevel(log.DebugLevel)
 		log.Warnf("已开启Debug模式.")
-		// log.Debugf("开发交流群: 192548878")
 	}
 	if !global.PathExists("device.json") {
 		log.Warn("虚拟设备信息不存在, 将自动生成随机设备.")
-		client.GenRandomDevice()
-		_ = os.WriteFile("device.json", client.SystemDeviceInfo.ToJson(), 0o644)
+		device = client.GenRandomDevice()
+		_ = os.WriteFile("device.json", device.ToJson(), 0o644)
 		log.Info("已生成设备信息并保存到 device.json 文件.")
 	} else {
 		log.Info("将使用 device.json 内的设备信息运行Bot.")
-		if err := client.SystemDeviceInfo.ReadJson([]byte(global.ReadAllText("device.json"))); err != nil {
+		device = new(client.DeviceInfo)
+		if err := device.ReadJson([]byte(global.ReadAllText("device.json"))); err != nil {
 			log.Fatalf("加载设备信息失败: %v", err)
 		}
+	}
+
+	if base.SignServer != "-" && base.SignServer != "" {
+		log.Infof("使用服务器 %s 进行数据包签名", base.SignServer)
+		wrapper.DandelionEnergy = energy
+		wrapper.FekitGetSign = sign
+	} else {
+		log.Warnf("警告: 未配置签名服务器, 这可能会导致登录 45 错误码或发送消息被风控")
 	}
 
 	if base.Account.Encrypt {
 		if !global.PathExists("password.encrypt") {
 			if base.Account.Password == "" {
 				log.Error("无法进行加密，请在配置文件中的添加密码后重新启动.")
-				readLine()
-				os.Exit(0)
+			} else {
+				log.Infof("密码加密已启用, 请输入Key对密码进行加密: (Enter 提交)")
+				byteKey, _ = term.ReadPassword(int(os.Stdin.Fd()))
+				base.PasswordHash = md5.Sum([]byte(base.Account.Password))
+				_ = os.WriteFile("password.encrypt", []byte(PasswordHashEncrypt(base.PasswordHash[:], byteKey)), 0o644)
+				log.Info("密码已加密，为了您的账号安全，请删除配置文件中的密码后重新启动.")
 			}
-			log.Infof("密码加密已启用, 请输入Key对密码进行加密: (Enter 提交)")
-			byteKey, _ = term.ReadPassword(int(os.Stdin.Fd()))
-			base.PasswordHash = md5.Sum([]byte(base.Account.Password))
-			_ = os.WriteFile("password.encrypt", []byte(PasswordHashEncrypt(base.PasswordHash[:], byteKey)), 0o644)
-			log.Info("密码已加密，为了您的账号安全，请删除配置文件中的密码后重新启动.")
 			readLine()
 			os.Exit(0)
-		} else {
-			if base.Account.Password != "" {
-				log.Error("密码已加密，为了您的账号安全，请删除配置文件中的密码后重新启动.")
-				readLine()
-				os.Exit(0)
-			}
-
-			if len(byteKey) == 0 {
-				log.Infof("密码加密已启用, 请输入Key对密码进行解密以继续: (Enter 提交)")
-				cancel := make(chan struct{}, 1)
-				state, _ := term.GetState(int(os.Stdin.Fd()))
-				go func() {
-					select {
-					case <-cancel:
-						return
-					case <-time.After(time.Second * 45):
-						log.Infof("解密key输入超时")
-						time.Sleep(3 * time.Second)
-						_ = term.Restore(int(os.Stdin.Fd()), state)
-						os.Exit(0)
-					}
-				}()
-				byteKey, _ = term.ReadPassword(int(os.Stdin.Fd()))
-				cancel <- struct{}{}
-			} else {
-				log.Infof("密码加密已启用, 使用运行时传递的参数进行解密，按 Ctrl+C 取消.")
-			}
-
-			encrypt, _ := os.ReadFile("password.encrypt")
-			ph, err := PasswordHashDecrypt(string(encrypt), byteKey)
-			if err != nil {
-				log.Fatalf("加密存储的密码损坏，请尝试重新配置密码")
-			}
-			copy(base.PasswordHash[:], ph)
 		}
+		if base.Account.Password != "" {
+			log.Error("密码已加密，为了您的账号安全，请删除配置文件中的密码后重新启动.")
+			readLine()
+			os.Exit(0)
+		}
+		if len(byteKey) == 0 {
+			log.Infof("密码加密已启用, 请输入Key对密码进行解密以继续: (Enter 提交)")
+			cancel := make(chan struct{}, 1)
+			state, _ := term.GetState(int(os.Stdin.Fd()))
+			go func() {
+				select {
+				case <-cancel:
+					return
+				case <-time.After(time.Second * 45):
+					log.Infof("解密key输入超时")
+					time.Sleep(3 * time.Second)
+					_ = term.Restore(int(os.Stdin.Fd()), state)
+					os.Exit(0)
+				}
+			}()
+			byteKey, _ = term.ReadPassword(int(os.Stdin.Fd()))
+			cancel <- struct{}{}
+		} else {
+			log.Infof("密码加密已启用, 使用运行时传递的参数进行解密，按 Ctrl+C 取消.")
+		}
+
+		encrypt, _ := os.ReadFile("password.encrypt")
+		ph, err := PasswordHashDecrypt(string(encrypt), byteKey)
+		if err != nil {
+			log.Fatalf("加密存储的密码损坏，请尝试重新配置密码")
+		}
+		copy(base.PasswordHash[:], ph)
 	} else if len(base.Account.Password) > 0 {
 		base.PasswordHash = md5.Sum([]byte(base.Account.Password))
 	}
@@ -202,10 +226,27 @@ func Main() {
 		time.Sleep(time.Second * 5)
 	}
 	log.Info("开始尝试登录并同步消息...")
-	log.Infof("使用协议: %s", client.SystemDeviceInfo.Protocol)
+	log.Infof("使用协议: %s", device.Protocol.Version())
 	cli = newClient()
+	cli.UseDevice(device)
 	isQRCodeLogin := (base.Account.Uin == 0 || len(base.Account.Password) == 0) && !base.Account.Encrypt
 	isTokenLogin := false
+
+	if isQRCodeLogin && cli.Device().Protocol != 2 {
+		log.Warn("当前协议不支持二维码登录, 请配置账号密码登录.")
+		os.Exit(0)
+	}
+
+	// 加载本地版本信息, 一般是在上次登录时保存的
+	versionFile := path.Join(global.VersionsPath, fmt.Sprint(int(cli.Device().Protocol))+".json")
+	if global.PathExists(versionFile) {
+		b, err := os.ReadFile(versionFile)
+		if err == nil {
+			_ = cli.Device().Protocol.Version().UpdateFromJson(b)
+		}
+		log.Infof("从文件 %s 读取协议版本 %v.", versionFile, cli.Device().Protocol.Version())
+	}
+
 	saveToken := func() {
 		base.AccountToken = cli.GenToken()
 		_ = os.WriteFile("session.token", base.AccountToken, 0o644)
@@ -220,8 +261,8 @@ func Main() {
 					log.Warnf("警告: 配置文件内的QQ号 (%v) 与缓存内的QQ号 (%v) 不相同", base.Account.Uin, cu)
 					log.Warnf("1. 使用会话缓存继续.")
 					log.Warnf("2. 删除会话缓存并重启.")
-					log.Warnf("请选择: (5秒后自动选1)")
-					text := readLineTimeout(time.Second*5, "1")
+					log.Warnf("请选择:")
+					text := readIfTTY("1")
 					if text == "2" {
 						_ = os.Remove("session.token")
 						log.Infof("缓存已删除.")
@@ -236,6 +277,7 @@ func Main() {
 				cli.Disconnect()
 				cli.Release()
 				cli = newClient()
+				cli.UseDevice(device)
 			} else {
 				isTokenLogin = true
 			}
@@ -244,6 +286,29 @@ func Main() {
 	if base.Account.Uin != 0 && base.PasswordHash != [16]byte{} {
 		cli.Uin = base.Account.Uin
 		cli.PasswordMd5 = base.PasswordHash
+	}
+	if !base.FastStart {
+		log.Infof("正在检查协议更新...")
+		currentVersionName := device.Protocol.Version().SortVersionName
+		remoteVersion, err := getRemoteLatestProtocolVersion(int(device.Protocol.Version().Protocol))
+		if err == nil {
+			remoteVersionName := gjson.GetBytes(remoteVersion, "sort_version_name").String()
+			if remoteVersionName != currentVersionName {
+				switch {
+				case !base.UpdateProtocol:
+					log.Infof("检测到协议更新: %s -> %s", currentVersionName, remoteVersionName)
+					log.Infof("如果登录时出现版本过低错误, 可尝试使用 -update-protocol 参数启动")
+				case !isTokenLogin:
+					_ = device.Protocol.Version().UpdateFromJson(remoteVersion)
+					log.Infof("协议版本已更新: %s -> %s", currentVersionName, remoteVersionName)
+				default:
+					log.Infof("检测到协议更新: %s -> %s", currentVersionName, remoteVersionName)
+					log.Infof("由于使用了会话缓存, 无法自动更新协议, 请删除缓存后重试")
+				}
+			}
+		} else if err.Error() != "remote version unavailable" {
+			log.Warnf("检查协议更新失败: %v", err)
+		}
 	}
 	if !isTokenLogin {
 		if !isQRCodeLogin {
@@ -308,6 +373,7 @@ func Main() {
 	})
 	saveToken()
 	cli.AllowSlider = true
+	download.SetTimeout(time.Duration(base.HTTPTimeout) * time.Second) // 在登录完成后设置, 防止在堵塞协议更新
 	log.Infof("登录成功 欢迎使用: %v", cli.Nickname)
 	log.Info("开始加载好友列表...")
 	global.Check(cli.ReloadFriendList(), true)
@@ -319,11 +385,16 @@ func Main() {
 		base.Account.Status = 0
 	}
 	cli.SetOnlineStatus(allowStatus[base.Account.Status])
-
 	servers.Run(coolq.NewQQBot(cli))
 	log.Info("资源初始化完成, 开始处理信息.")
 	log.Info("アトリは、高性能ですから!")
+}
 
+// WaitSignal 在新线程检查更新和网络并等待信号, 必须在 InitBase, PrepareData, LoginInteract 之后执行
+//
+//   - 直接返回: os.Interrupt, syscall.SIGTERM
+//   - dump stack: syscall.SIGQUIT, syscall.SIGUSR1
+func WaitSignal() {
 	go func() {
 		selfupdate.CheckUpdate()
 		selfdiagnosis.NetworkDiagnosis(cli)
@@ -384,6 +455,23 @@ func newClient() *client.QQClient {
 	}
 	c.SetLogger(protocolLogger{})
 	return c
+}
+
+var remoteVersions = map[int]string{
+	1: "https://raw.githubusercontent.com/RomiChan/protocol-versions/master/android_phone.json",
+	6: "https://raw.githubusercontent.com/RomiChan/protocol-versions/master/android_pad.json",
+}
+
+func getRemoteLatestProtocolVersion(protocolType int) ([]byte, error) {
+	url, ok := remoteVersions[protocolType]
+	if !ok {
+		return nil, errors.New("remote version unavailable")
+	}
+	response, err := download.Request{URL: url}.Bytes()
+	if err != nil {
+		return download.Request{URL: "https://ghproxy.com/" + url}.Bytes()
+	}
+	return response, nil
 }
 
 type protocolLogger struct{}
