@@ -4,10 +4,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/Mrs4s/MiraiGo/topic"
-
 	"github.com/Mrs4s/MiraiGo/client"
 	"github.com/Mrs4s/MiraiGo/message"
+	"github.com/Mrs4s/MiraiGo/topic"
 	log "github.com/sirupsen/logrus"
 
 	"github.com/Mrs4s/go-cqhttp/global"
@@ -71,13 +70,12 @@ func (bot *CQBot) formatGroupMessage(m *message.GroupMessage) *event {
 		typ = "message_sent/group/normal"
 	}
 	gm := global.MSG{
-		"anonymous":    nil,
-		"font":         0,
-		"group_id":     m.GroupCode,
-		"message":      ToFormattedMessage(m.Elements, source),
-		"message_type": "group",
-		"message_seq":  m.Id,
-		"raw_message":  cqm,
+		"anonymous":   nil,
+		"font":        0,
+		"group_id":    m.GroupCode,
+		"message":     ToFormattedMessage(m.Elements, source),
+		"message_seq": m.Id,
+		"raw_message": cqm,
 		"sender": global.MSG{
 			"age":     0,
 			"area":    "",
@@ -94,7 +92,7 @@ func (bot *CQBot) formatGroupMessage(m *message.GroupMessage) *event {
 			"name": m.Sender.AnonymousInfo.AnonymousNick,
 		}
 		gm["sender"].(global.MSG)["nickname"] = "匿名消息"
-		gm["sub_type"] = "anonymous"
+		typ = "message/group/anonymous"
 	} else {
 		group := bot.Client.FindGroup(m.GroupCode)
 		mem := group.FindMember(m.Sender.Uin)
