@@ -13,9 +13,11 @@ import (
 	"strings"
 
 	"github.com/Mrs4s/MiraiGo/utils"
+	b14 "github.com/fumiama/go-base16384"
+	"github.com/segmentio/asm/base64"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/Mrs4s/go-cqhttp/internal/param"
+	"github.com/Mrs4s/go-cqhttp/internal/download"
 )
 
 const (
@@ -25,6 +27,8 @@ const (
 	VoicePath = "data/voices"
 	// VideoPath go-cqhttp使用的视频缓存目录
 	VideoPath = "data/videos"
+	// VersionsPath go-cqhttp使用的版本信息目录
+	VersionsPath = "data/versions"
 	// CachePath go-cqhttp使用的缓存目录
 	CachePath = "data/cache"
 	// DumpsPath go-cqhttp使用错误转储目录
@@ -82,16 +86,22 @@ func FindFile(file, cache, p string) (data []byte, err error) {
 		if (cache == "" || cache == "1") && PathExists(cacheFile) {
 			return os.ReadFile(cacheFile)
 		}
-		data, err = GetBytes(file)
-		_ = os.WriteFile(cacheFile, data, 0o644)
+		err = download.Request{URL: file}.WriteToFile(cacheFile)
 		if err != nil {
 			return nil, err
 		}
+		return os.ReadFile(cacheFile)
 	case strings.HasPrefix(file, "base64"):
-		data, err = param.Base64DecodeString(strings.TrimPrefix(file, "base64://"))
+		data, err = base64.StdEncoding.DecodeString(strings.TrimPrefix(file, "base64://"))
 		if err != nil {
 			return nil, err
 		}
+	case strings.HasPrefix(file, "base16384"):
+		data, err = b14.UTF82UTF16BE(utils.S2B(strings.TrimPrefix(file, "base16384://")))
+		if err != nil {
+			return nil, err
+		}
+		data = b14.Decode(data)
 	case strings.HasPrefix(file, "file"):
 		var fu *url.URL
 		fu, err = url.Parse(file)
